Add X in required encodings for bubble rank to enable labels!
if 'x' added to requiredEncodings then we have 'label' in frames
if requiredEncodings is empty then we have fulljoin for encoding data in dataMapCache. but 'label' data is 'lookups' dataframe which does not have coded iterators yet ). we have warning "Invalid operation. Generated dataframe iterator not yet implemented." and miss 'label' in datamapcache.
https://github.com/vizabi/vizabi-reactive/blob/master/src/core/marker/baseMarker.js#L146
with 'x' in requiredEncodings all another encoding (non defining data) have leftjoin and it worked